DT-610-3: Split end point into two
