+++ auth-service/src/main/kotlin/uk/gov/communities/delta/auth/Injection.kt
@@ -284,6 +284,10 @@ class Injection(
         userLookupService,
         userService,
         groupService,
+    )
+
+    fun adminGetUserController() = AdminGetUserController(
+        userLookupService,
         organisationService,
         accessGroupsService,
         ::MemberOfToDeltaRolesMapper
