+++ auth-service/src/test/kotlin/uk/gov/communities/delta/controllers/AdminEmailControllerTest.kt
@@ -312,6 +312,7 @@ class AdminEmailControllerTest {
                             mockk(relaxed = true),
                             mockk(relaxed = true),
                             mockk(relaxed = true),
+                            mockk(relaxed = true),
                         )
                     }
                 }
