+++ auth-service/src/test/kotlin/uk/gov/communities/delta/controllers/AdminUserCreationControllerTest.kt
@@ -454,6 +454,7 @@ class AdminUserCreationControllerTest {
                             mockk(relaxed = true),
                             controller,
                             mockk(relaxed = true),
+                            mockk(relaxed = true),
                         )
                     }
                 }
