+++ auth-service/src/test/kotlin/uk/gov/communities/delta/controllers/FetchUserAuditControllerTest.kt
@@ -201,6 +201,7 @@ class FetchUserAuditControllerTest {
                             controller,
                             mockk<AdminUserCreationController>(relaxed = true),
                             mockk(relaxed = true),
+                            mockk(relaxed = true),
                         )
                     }
                 }
