+++ auth-service/src/test/kotlin/uk/gov/communities/delta/controllers/RefreshUserInfoControllerTest.kt
@@ -131,6 +131,7 @@ class RefreshUserInfoControllerTest {
                             mockk<FetchUserAuditController>(relaxed = true),
                             mockk<AdminUserCreationController>(relaxed = true),
                             mockk(relaxed = true),
+                            mockk(relaxed = true),
                         )
                     }
                 }
